fix: renamed the OnboardingReport file case
