+++ kits/codebase-onboarding-agent/apps/src/components/OnboardingReport.tsx
